tweak & refactor branch widget display
[] is now used for branches that have a corresponding remote branch.
() is now used for branches that don't.

x is now used for merges that occur locally & remotely
~ is now used for merges that occur locally only

Note that () used to be used for merges that only occur locally. So this
is a potentially confusing change, but I think it gives a little more
information in the same amount of space.