show remote branches on origin too
This is a big changeset. The idea is that we should be showing remote
branches on origin, too, even if they don't have a local version, since
that's often the primary point point of synchronization.

So now we have remote-only branches, denoted by { }, and lots of things
have to change to accomodate that.